Replace image details page with new one
