+++ src/components/VImageDetails/VRelatedImages.vue
@@ -3,8 +3,10 @@
     <h3 class="text-2xl md:text-3xl mb-6">
       {{ $t('image-details.related-images') }}
     </h3>
+    <GridSkeleton v-if="!imageId" is-for-tab="image" :nums-elems="10" />
     <!-- TODO: use a image grid with correct margins and items gaps -->
     <!-- <ImageGrid
+      v-if="imageId"
       :images="images"
       :can-load-more="false"
       :fetch-state="{
@@ -16,29 +18,13 @@
 </template>
 
 <script>
-import { ref } from '@nuxtjs/composition-api'
-import useRelated from '~/composables/use-related'
-import { IMAGE } from '~/constants/media'
-// import ImageGrid from '~/components/ImageGrid/ImageGrid'
-
 export default {
   name: 'VRelatedImages',
   // components: { ImageGrid },
   props: {
     imageId: {
       type: String,
-      required: true,
     },
-  },
-  setup(props) {
-    const mainImageId = ref(props.imageId)
-    const relatedOptions = {
-      mediaType: IMAGE,
-      mediaId: mainImageId,
-    }
-    const { media: images } = useRelated(relatedOptions)
-
-    return { images }
   },
 }
 </script>
